+++ src/sound/ResampleTrivial.hh
@@ -11,7 +11,7 @@ class ResampleTrivial final : public ResampleAlgo
 {
 public:
 	explicit ResampleTrivial(ResampledSoundDevice& input);
-	bool generateOutputImpl(float* dataOut, unsigned num,
+	bool generateOutputImpl(float* dataOut, size_t num,
 	                        EmuTime::param time) override;
 };
